+++ src/main/kotlin/cloud/glitchdev/rfu/events/managers/ChatEvents.kt
@@ -7,6 +7,7 @@ import cloud.glitchdev.rfu.utils.RFULogger
 import gg.essential.universal.utils.toFormattedString
 import gg.essential.universal.utils.toUnformattedString
 import net.fabricmc.fabric.api.client.message.v1.ClientReceiveMessageEvents
+import net.fabricmc.fabric.api.client.message.v1.ClientSendMessageEvents
 import net.minecraft.network.chat.Component
 
 @AutoRegister
@@ -23,6 +24,32 @@ object ChatEvents : RegisteredEvent {
             if(!result) RFULogger.dev("Game message was hid: ${message.toFormattedString()}")
             result
         }
+
+        ClientSendMessageEvents.ALLOW_CHAT.register { message ->
+            val result = SendChatEventManager.runTasks(message)
+            if(!result) RFULogger.dev("Sent message was cancelled: $message")
+            result
+        }
+
+        ClientSendMessageEvents.ALLOW_COMMAND.register { command ->
+            val result = SendCommandEventManager.runTasks(command)
+            if(!result) RFULogger.dev("Sent commnad was cancelled: $command")
+            result
+        }
+    }
+
+    fun registerSendChatEvent(
+        priority: Int = 20,
+        callback: (message: String) -> Boolean
+    ) : SendChatEventManager.SendChatEvent {
+        return SendChatEventManager.register(priority, callback)
+    }
+
+    fun registerSendCommandEvent(
+        priority: Int = 20,
+        callback: (message: String) -> Boolean
+    ) : SendCommandEventManager.SendCommandEvent {
+        return SendCommandEventManager.register(priority, callback)
     }
 
     fun registerAnyChatEvent(
@@ -131,6 +158,50 @@ object ChatEvents : RegisteredEvent {
             priority: Int = 20,
             callback: (text: Component, overlay: Boolean) -> Boolean
         ) : ManagedTask<(text: Component, overlay: Boolean) -> Boolean, GameEvent>(priority, callback) {
+            override fun register() = submitTask(this)
+            override fun unregister() = removeTask(this)
+        }
+    }
+
+    object SendChatEventManager : AbstractEventManager<(message: String) -> Boolean, SendChatEventManager.SendChatEvent>() {
+        override val runTasks: (message: String) -> Boolean = { message ->
+            var result = true
+            safeExecution {
+                for (event in tasks) if (!event.callback(message)) result = false
+            }
+            result
+        }
+
+        fun register(priority: Int = 20, callback: (message: String) -> Boolean): SendChatEvent {
+            return SendChatEvent(priority, callback).register()
+        }
+
+        class SendChatEvent(
+            priority: Int = 20,
+            callback: (message: String) -> Boolean
+        ) : ManagedTask<(message: String) -> Boolean, SendChatEvent>(priority, callback) {
+            override fun register() = submitTask(this)
+            override fun unregister() = removeTask(this)
+        }
+    }
+
+    object SendCommandEventManager : AbstractEventManager<(command: String) -> Boolean, SendCommandEventManager.SendCommandEvent>() {
+        override val runTasks: (command: String) -> Boolean = { command ->
+            var result = true
+            safeExecution {
+                for (event in tasks) if (!event.callback(command)) result = false
+            }
+            result
+        }
+
+        fun register(priority: Int = 20, callback: (command: String) -> Boolean): SendCommandEvent {
+            return SendCommandEvent(priority, callback).register()
+        }
+
+        class SendCommandEvent(
+            priority: Int = 20,
+            callback: (command: String) -> Boolean
+        ) : ManagedTask<(command: String) -> Boolean, SendCommandEvent>(priority, callback) {
             override fun register() = submitTask(this)
             override fun unregister() = removeTask(this)
         }